Move Windows-specific code to win32 and change 'audio' to 'sound'
